pyproject.toml
[project]
name = "armada_jupyter"
version = "0.1.0"
description = "Submit JupyterLab Pods to Armada"
readme = "README.md"
requires-python = ">=3.7,<3.9"
dependencies = ["armada_client", "typer", "pyyaml", "types-PyYAML"]
license = { text = "Apache Software License" }
authors = [{ name = "G-Research Open Source Software", email = "armada@armadaproject.io" }]

[project.optional-dependencies]
format = ["black~=22.3.0", "flake8~=4.0.1", "pylint~=2.13.8", "mypy~=0.971" ]
test = ["pytest~=7.1.2", "mock~=4.0.3"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
filterwarnings = [
    'ignore: the imp module is deprecated:DeprecationWarning',
]